extract ShareSequence and rename share_merging to parse
