+++ pkg/shares/share_sequence.go
@@ -1,90 +1,18 @@
 package shares
 
 import (
-	"bytes"
 	"fmt"
 
 	"github.com/celestiaorg/celestia-app/pkg/appconsts"
 	"github.com/celestiaorg/nmt/namespace"
-	coretypes "github.com/tendermint/tendermint/types"
 )
-
-// ParseTxs collects all of the transactions from the shares provided
-func ParseTxs(shares [][]byte) (coretypes.Txs, error) {
-	// parse the sharse
-	rawTxs, err := parseCompactShares(shares, appconsts.SupportedShareVersions)
-	if err != nil {
-		return nil, err
-	}
-
-	// convert to the Tx type
-	txs := make(coretypes.Txs, len(rawTxs))
-	for i := 0; i < len(txs); i++ {
-		txs[i] = coretypes.Tx(rawTxs[i])
-	}
-
-	return txs, nil
-}
-
-// ParseBlobs collects all blobs from the shares provided
-func ParseBlobs(shares [][]byte) ([]coretypes.Blob, error) {
-	blobList, err := parseSparseShares(shares, appconsts.SupportedShareVersions)
-	if err != nil {
-		return []coretypes.Blob{}, err
-	}
-
-	return blobList, nil
-}
 
 // ShareSequence represents a contiguous sequence of shares that are part of the
 // same namespace and blob. For compact shares, one share sequence exists per
 // reserved namespace. For sparse shares, one share sequence exists per blob.
-// TODO consider extracting the ShareSequence struct to a new file.
 type ShareSequence struct {
 	NamespaceID namespace.ID
 	Shares      []Share
-}
-
-func ParseShares(rawShares [][]byte) ([]ShareSequence, error) {
-	sequences := []ShareSequence{}
-	currentSequence := ShareSequence{}
-
-	for _, rawShare := range rawShares {
-		share, err := NewShare(rawShare)
-		if err != nil {
-			return sequences, err
-		}
-		isStart, err := share.IsSequenceStart()
-		if err != nil {
-			return sequences, err
-		}
-		if isStart {
-			if len(currentSequence.Shares) > 0 {
-				sequences = append(sequences, currentSequence)
-			}
-			currentSequence = ShareSequence{
-				Shares:      []Share{share},
-				NamespaceID: share.NamespaceID(),
-			}
-		} else {
-			if !bytes.Equal(currentSequence.NamespaceID, share.NamespaceID()) {
-				return sequences, fmt.Errorf("share sequence %v has inconsistent namespace IDs with share %v", currentSequence, share)
-			}
-			currentSequence.Shares = append(currentSequence.Shares, share)
-		}
-	}
-
-	if len(currentSequence.Shares) > 0 {
-		sequences = append(sequences, currentSequence)
-	}
-
-	for _, sequence := range sequences {
-		if err := sequence.validSequenceLen(); err != nil {
-			return sequences, err
-		}
-	}
-
-	return sequences, nil
 }
 
 // RawData returns the raw share data of this share sequence. The raw data does
@@ -106,6 +34,14 @@ func (s ShareSequence) RawData() (data []byte, err error) {
 	return data[:sequenceLen], nil
 }
 
+func (s ShareSequence) SequenceLen() (uint32, error) {
+	if len(s.Shares) == 0 {
+		return 0, fmt.Errorf("invalid sequence length because share sequence %v has no shares", s)
+	}
+	firstShare := s.Shares[0]
+	return firstShare.SequenceLen()
+}
+
 // validSequenceLen extracts the sequenceLen written to the first share
 // and returns an error if the number of shares needed to store a sequence of
 // length sequenceLen doesn't match the number of shares in this share
@@ -124,14 +60,6 @@ func (s ShareSequence) validSequenceLen() error {
 		return fmt.Errorf("share sequence has %d shares but needed %d shares", len(s.Shares), sharesNeeded)
 	}
 	return nil
-}
-
-func (s ShareSequence) SequenceLen() (uint32, error) {
-	if len(s.Shares) == 0 {
-		return 0, fmt.Errorf("invalid sequence length because share sequence %v has no shares", s)
-	}
-	firstShare := s.Shares[0]
-	return firstShare.SequenceLen()
 }
 
 // numberOfSharesNeeded extracts the sequenceLen written to the share
